Date: October 21, 2025
Dear Dexter,

I apologize for how late this is being sent. I can completely understanding if you cannot look at this if you are currently on parental leave (since its post October 1st) if plans remained the same or because of the government shutdown.

General Overview:
All aspects of the data wrangling and data analysis concerning landscape mullets have been completed—with me taking as many steps to try to account for potential errors.

0_data_wrangling Overview:
As discussed in the summer, all datasets were consolidated and aggregated into one script, which has all of the
different items considered:

1. property parcels
2. census data
3. iTree
4. general cleaning
5. combination.

Y2019 and Y2023 parcel data was used since it was the most consistent between Holyoke and Chelsea (they are assessed at different amounts each year); also, made sure to include inflation into account. Census data was cleaned a bit more extensively from last time and includes percentages for easier comparisons—the committ sends those respective geopackages. iTree remains the same but for general cleaning, I calculated the Days in the Ground using your previous code and technique, with '1" being event triggers (i.e. death) and "0" remaining the same—they had survey-date for 2017 and 2024 (at least for Holyoke), which is how that was estimated. The final geopackage was exported with all of these variables accounted for

1_analysis Overview:
Major takeaway is that most of the statistical test behind FY and BY is insignifcant due to such a small sample size: conducted three corresponding chi squares/fisher test (dependent on event amount) and three survivorship curves given your code. Currently, unsure what to do because of this. However, an idea is potentially to increase the level of my question—how does parcel value, building style, type of tree (ornamental), census,itree impact survivorship (since I have the days in the ground and the time-long data already). I will likely meet with Prof. Grove soon to be able to flesh this out quickly to finish that aspect of the analysis.

Thank you again for your time and guidance throughout this process! Please let me know if there’s a better time to discuss this once you’re available but completely understand if not.

Best,
Eduardo