Date: November 13th, 2025 Summary: 00_wrangling_new: Fixed the wrangling script with the changes of census data, inflating property values, and other minute differences. Additionally, downloaded and started using the dataset given from HERO. There are some discrepancies with the NAs there and the presentaiton (although, a lot less :D ). Fixed my code to work with the new dataset.
01_data_analysis_new:
Fixed Chi Squares from our last conversation to observe both the overall dataset and the residential-only dataset. Organized it in a way to be esasily interpretable. Secondly, did the same approach from the KM curves and decided whether to use a variable from the three overaching categories—Ecological, Parcel, and Census—based on this. I have an explanation and some notes for each of those conducted. Finally, performed the CH Proporational Model for each category and created visualizations for each of them. The final conclusion seemed to be that Ecological variables were the best predictors; however, I fear that some of the other categories were affected by colinearity and overfitting (so, this likely needs review).

Next Steps:
Unsure what they are, exactly. It might be making the "mega model" but need your advice if what I did thus far is correct. This is still very interesting and am excited to write these results down after a double check. Also, there are places within both scripts that are "ESM to DHL" which are comments directed to you, usually for review. Please let me know if there is anything confusing (sorry for taking a bit with the code!)